Port to GCC 6.1.1 as found on Fedora 24.  This is most of the way there, but not everything compiles yet.
g++ now adds an implicit throw declaration to destructors based on what
member destructors may throw, and it enforces that subclass implementations
of virtual destructors must not have looser exceptions, and if nothing in
the base class destructor might throw an exception, or the base class
destructor is explicitly declared not to throw an exception, then the
subclass implementations cannot declare an exception to throw either,
whether implicit or explicit

  http://stackoverflow.com/questions/11497252/default-destructor-nothrow

If the throw declaration is removed from the members' destructors, then you
get this warning:

nidas/util/ThreadSupport.cc:256:55: warning: throw will always call terminate() [-Wterminate]
             throw Exception("~Mutex","Mutex is locked");
                                                       ^
nidas/util/ThreadSupport.cc:256:55: note: in C++11 destructors default to noexcept

So instead, it seems better to just expand the throw declaration for the
destructor in the base class.  That at least matches past actual behavior.

auto_ptr is deprecated and now produces many verbose warnings.  Process()
makes heavy use of it, and since the auto_ptr copy constructor and
assignment semantics were to transfer ownership, auto_ptr should be
replaced with unique_ptr.  In C++11 unique_ptr has an explicit assignment
operator which transfers ownership, and the copy constructor signature is
deleted, but default gcc 6.1.1 (without specifying the C++11 features) does
not have the assignment operator.  So any auto_ptr copies or assignments
have to be replaced with newowner.reset(oldowner.release()).

make_pair<> should not be instantiated with types, it is a helper function
and should be allowed to deduce types itself.  Now that make_pair<> wants
to use &&, the explicit types in all the uses of make_pair() need to be
removed.

Static const non-integers cannot be initialized in the class definition
without constexprt, but constexpr is not necessarily available on older
versions.  So such constants need to be defined as static const in the
class without initializing it, then it must be defined in the source module
same as other static class members.