try adding explicit exception specifiers for destructors
Then there's the issue with destructors between 4.8.5 and 6.1.1.  It looks
like destructors in 4.8.5 are by default open exception specs, meaning if a
base class declares an exception specifier, but then a subclass does not,
then the subclass is considered to be loosening the exceptions which might
be thrown.  One fix appears to be to make the exception specifier explicit
on all subclasses, but that's a lot of classes in nidas.  This change is a
start on that for testing against 6.1.1, but I suspect the better way is to
make sure no destructors can throw exceptions, and then leave the exception
specifiers off and accept the compiler default.